Improved code for XYplorer window detection and fixed saving of configuration.
The detection of XYplorer's window is now much better and should find
the proper and correct XYplorer window to send the XYplorer's script to.
Also fixed the issue where the applications that are checked to allow
FolderMenu3 EX to show up in were saved with a value of "-1" in the
config file when it's supposed to be "1" for activated and "0" for
non-activated.